+++ magda-registry-api/src/main/scala/au/csiro/data61/magda/registry/RecordPersistence.scala
@@ -63,6 +63,10 @@ trait RecordPersistence {
       aspectOrQueries: Iterable[AspectQuery] = Nil
   )(implicit session: DBSession): Long
 
+  def getReadiness(
+      logger: Option[LoggingAdapter] = None
+  )(implicit session: DBSession): Option[Boolean]
+
   def getById(
       tenantId: TenantId,
       opaQueries: Option[List[(String, List[List[OpaQuery]])]],
@@ -245,6 +249,7 @@ class DefaultRecordPersistence(config: Config)
 
   /**
     * Given a list of recordIds, filter out any record that the current user has not access and return the rest of ids
+    *
     * @param tenantId
     * @param opaRecordQueries
     * @param recordIds
@@ -321,6 +326,51 @@ class DefaultRecordPersistence(config: Config)
         Some(SQLSyntax.roundBracket(sqlPart))
       case _ => None
     }
+  }
+
+  def getReadiness(
+      logger: Option[LoggingAdapter] = None
+  )(implicit session: DBSession): Option[Boolean] = {
+    def checkTableExists(table: String): Boolean = {
+      try {
+        val schemaAndTable = "public." + table
+        val regclass = sql"select to_regclass(${schemaAndTable})"
+
+        val tableExists = regclass
+          .map(_.toMap)
+          .list
+          .apply()
+          .map { res =>
+            {
+              res("to_regclass").toString == table
+            }
+          }
+          .forall(x => x);
+        tableExists
+      } catch {
+        case e: Exception => {
+          if (logger.isDefined) {
+            logger.get.error(
+              e,
+              s"Error occurred on getReadiness probe for table ${table}"
+            )
+          }
+          false
+        }
+      }
+    }
+
+    Some(
+      List(
+        checkTableExists("aspects"),
+        checkTableExists("events"),
+        checkTableExists("eventtypes"),
+        checkTableExists("recordaspects"),
+        checkTableExists("records"),
+        checkTableExists("webhookevents"),
+        checkTableExists("webhooks")
+      ).forall(x => x)
+    )
   }
 
   def getAllWithAspects(
@@ -676,7 +726,9 @@ where (RecordAspects.recordId, RecordAspects.aspectId)=($recordId, $aspectId) AN
         case _                           => Success(result)
       }
       // No failed aspects, so unwrap the aspects from the Success Trys.
-      resultAspects <- Try { patchedAspects.mapValues(_.get) }
+      resultAspects <- Try {
+        patchedAspects.mapValues(_.get)
+      }
     } yield
       result.copy(aspects = resultAspects, sourceTag = newRecord.sourceTag)
   }
@@ -1368,7 +1420,7 @@ where (RecordAspects.recordId, RecordAspects.aspectId)=($recordId, $aspectId) AN
       aspectIds: Iterable[String] = List()
   )(implicit session: DBSession): Try[Iterable[String]] = {
 
-    /** For aspects that have links to other aspects, a map of the ids of those aspects to the location (first-level, not path) in their JSON where the link is  */
+    /** For aspects that have links to other aspects, a map of the ids of those aspects to the location (first-level, not path) in their JSON where the link is */
     val referenceMap = this.buildReferenceMap(aspectIds)
 
     Try {
@@ -1952,6 +2004,7 @@ where (RecordAspects.recordId, RecordAspects.aspectId)=($recordId, $aspectId) AN
       case _            => throw new Exception("Invalid tenant value " + tenantId)
     }
   }
+
   private def aspectIdsToWhereClause(
       tenantId: TenantId,
       aspectIds: Iterable[String]